requirements.txt
fastapi==0.68.0
pydantic==1.8.2
cachetools==4.2.4
uvicorn==0.23.2
pyjwt==2.8.0